[core] AbstractFileStoreWrite should not be the interface
